+++ src/ForWebforms.Tests/Expressions/Convention.cs
@@ -1,0 +1,6 @@
+﻿namespace ExitStrategy.ForWebforms.Tests.Expressions
+{
+    public class Convention : BasicConvention
+    {
+    }
+}
